Rename header and footer files and components without creating name clash with native base components
